+++ include/vlink/internal/getter-inl.h
@@ -148,12 +148,6 @@ inline void Getter<ValueT, SecT>::set_change_reporting(bool enable) {
 }
 
 template <typename ValueT, SecurityType SecT>
-inline void Getter<ValueT, SecT>::set_manual_unloan(bool manual_unloan) {
-  this->impl_->set_manual_unloan(manual_unloan);
-  this->is_manual_unloan_ = manual_unloan;
-}
-
-template <typename ValueT, SecurityType SecT>
 inline void Getter<ValueT, SecT>::set_latency_and_lost_enabled(bool enable) {
   this->impl_->set_latency_and_lost_enabled(enable);
 }