+++ modules/shm2/shm2_factory.h
@@ -433,10 +433,6 @@ class Shm2Subscriber final : public AbstractObject<ShmID2>, public std::enable_s
 
   void unsubscribe();
 
-  void set_manual_unloan(bool manual_unloan);
-
-  bool release(const Bytes& bytes);
-
   void set_latency_and_lost_enabled(bool enable);
 
   bool is_latency_and_lost_enabled() const;
@@ -471,15 +467,6 @@ class Shm2Subscriber final : public AbstractObject<ShmID2>, public std::enable_s
 
   std::atomic_bool is_latency_and_lost_enabled_{false};
   CalculateSample calc_sample_;
-  std::atomic_bool manual_unloan_{false};
-
-  struct SubscriberLoanEntry {
-    std::unique_ptr<iox2_sample_t> storage;
-    iox2_sample_h handle{nullptr};
-  };
-
-  std::mutex loan_mtx_;
-  std::unordered_map<const uint8_t*, SubscriberLoanEntry> loan_map_;
 };
 
 }  // namespace vlink